pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "botmusica"
version = "0.1.0"
description = "Bot de musica para Discord com comandos slash"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py[voice]>=2.4.0,<3.0.0",
  "PyNaCl>=1.5.0,<2.0.0",
  "yt-dlp>=2025.1.0",
  "wavelink>=3.4.1,<4.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
]
postgres = [
  "psycopg[binary]>=3.2.0,<4.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
